requirements.txt
# appdirs==1.4.4
# appnope==0.1.2
# arrow==1.1.1
# asn1crypto==1.4.0
# backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.3
# blinker==1.4
# cffi==1.14.6
# click==8.0.1
cssmin==0.2.0
# docutils==0.17.1
feedgenerator==1.9.1
# idna==3.2
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
pelican==4.6.0
# pycparser==2.20
Pygments==2.9.0
pyScss==1.3.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
# Unidecode==1.2.0
webassets==2.0
